// src/app/types/PatientData.ts
export interface PatientData {
    patientId: string;
    patientId1: string;
    firstName: string;
    lastName: string;
    email: string;
    phone: number;
    address: string;
    city: string;
    state: string;
    zip: string;
    bloodGroup: string;
    dateOfBirth: Date;
    gender: string;
    pharmacyId: string;
 }